Fix tests for linux
