+++ Tests/PlayingCardTests/SuitTests.swift
@@ -27,4 +27,8 @@ class SuitTests: XCTestCase {
         XCTAssertGreaterThan(Suit.diamonds, Suit.clubs)
     }
     
+    static var allTests = [
+        ("testSuitStringEquality", testSuitStringEquality),
+        ("testSuitComparable", testSuitComparable),
+    ]
 }
